+++ src/main/java/com/project/cookaround/domain/member/service/MemberService.java
@@ -3,16 +3,10 @@ package com.project.cookaround.domain.member.service;
 import com.project.cookaround.domain.member.entity.Member;
 import com.project.cookaround.domain.member.repository.MemberRepository;
 import jakarta.transaction.Transactional;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
-
 @Service
-public class MemberService implements UserDetailsService {
+public class MemberService {
 
     private final MemberRepository memberRepository;
 
@@ -35,17 +29,6 @@ public class MemberService implements UserDetailsService {
 
     public boolean validateDuplicateMemberByEmail(String email) {
         return memberRepository.findByEmail(email).isPresent();
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String loginId) throws UsernameNotFoundException {
-        Optional<Member> member = Optional.ofNullable(memberRepository.findByLoginId(loginId)
-                .orElseThrow(() -> new UsernameNotFoundException(loginId)));
-
-        return User.builder()
-                .username(member.get().getLoginId())
-                .password(member.get().getPassword())
-                .build();
     }
 
 }
